requirements.txt
anthropic
tavily-python
python-telegram-bot
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
fastapi
uvicorn[standard]
pydantic